requirements.txt
Nuitka==2.7
ordered-set==4.1.0
pyaml==25.1.0
python-dotenv==1.1.0
PyYAML==6.0.2
zstandard==0.23.0
